[SDRAN-1480] Adding ArtifactName property so that artifacts can be named differently than the model itself
This is needed as in some cases we have different versions of the model coexisting at the same time, eg 'aether-2.x` and `aether-4.0.x`.
It is not enough to set a different version of the artifact as we want to be able to generate a Docker container with tag 'latest' for both of them.